requirements.txt
python-dotenv
selenium
google-generativeai
webdriver-manager